prescription_ocr/requirements.txt
opencv-python
pillow
numpy
pandas
matplotlib
scikit-learn
tensorflow>=2.9.0
tqdm
pytesseract
jiwer